requirements.txt
openai==1.55.3
langchain==0.2.7
langchain-community==0.2.7
langchain-huggingface==0.0.3
langchain-experimental==0.0.62
langchain-openai==0.1.14
chromadb==0.5.3
sentence-transformers==3.0.1
python-dotenv==1.0.1
lark==1.1.9
llama-index-core
llama-parse==0.5.11
llama-index-readers-file
llama-index-llms-langchain
langgraph
groq
azure-storage-blob
pypdf
sentence-transformers
pinecone-client
python-dotenv
azure-storage-blob